+++ routes/classes.js
@@ -1,0 +1,5 @@
+module.exports = function (app) {
+    app.get('/classes', (req, res) => {
+        res.render("classes.ejs");
+    });
+};
