--- routes/index.js
@@ -1,10 +1,0 @@
-app.get('/login', (req, res) => {
-    if (req.query.token) {
-        let tokenData = jwt.decode(req.query.token)
-        req.session.token = tokenData
-        req.session.user = tokenData.displayName
-        res.redirect('/')
-    } else {
-        res.redirect(`${FBJS_URL}/oauth?redirectURL=${THIS_URL}`)
-    }
-})
