+++ routes/queues.js
@@ -1,3 +1,7 @@
+module.exports = function (app) {
+
 app.get('/queue', (req, res) => {
     res.render('queue.ejs')
 })
+
+};
